+++ packages/framework/bundle/event-map/bundle.js
@@ -56,22 +56,7 @@ function() {
 	"protocol" : ["Oskari.bundle.Bundle"],
 	"source" : {
 
-		"scripts" : [
-		/*
-		 * map base (includes layer basics, map basics )
-		 */
-
-		{
-			"type" : "text/javascript",
-			"src" : "../../../../sources/framework/event/common/after-map-layer-add-event.js"
-		}, {
-			"type" : "text/javascript",
-			"src" : "../../../../sources/framework/event/common/after-map-layer-remove-event.js"
-		}, {
-			"type" : "text/javascript",
-			"src" : "../../../../sources/framework/event/common/MapLayerEvent.js"
-		}
-        ],
+		"scripts" : [],
 		"resources" : []
 	},
 	"bundle" : {