+++ packages/framework/bundle/sandbox-base/bundle.js
@@ -61,9 +61,6 @@ function() {
 			"src" : "../../../../sources/framework/sandbox/sandbox.js"
 		}, {
 			"type" : "text/javascript",
-			"src" : "../../../../sources/framework/sandbox/sandbox-key-listener-methods.js"
-		}, {
-			"type" : "text/javascript",
 			"src" : "../../../../sources/framework/sandbox/sandbox-state-methods.js"
 		}],
 		"resources" : []