+++ sources/framework/domain/map.js
@@ -8,14 +8,18 @@
  * way to control the map. This is only to get map values without openlayers
  * dependency.
  */
-Oskari.clazz.define('Oskari.mapframework.domain.Map',
+(function(Oskari) {
+    var log = Oskari.log('map.state');
 
-    /** 
+    Oskari.clazz.define('Oskari.mapframework.domain.Map',
+
+    /**
      * @method create called automatically on construction
      * @static
      */
 
-    function () {
+    function (sandbox) {
+        this._sandbox = sandbox;
 
         // @property {Number} _centerX map longitude (float)
         this._centerX = null;
@@ -50,11 +54,19 @@ Oskari.clazz.define('Oskari.mapframework.domain.Map',
         // @property {OpenLayers.Bounds} maximumExtent configured for the map { left: NaN, top: NaN, right: NaN, bottom: NaN }
         this.maxExtent = null;
 
-        // @property {Boolean} _isMoving true when map is being dragged 
+        // @property {Boolean} _isMoving true when map is being dragged
         this._isMoving = false;
 
         // @property {String} _projectionCode SRS projection code, defaults to 'EPSG:3067'
         this._projectionCode = "EPSG:3067";
+
+        // moved from core
+        this._selectedLayers = [];
+
+        // moved from core "highlighted" layers
+        this._activatedLayers = [];
+
+        this._allowMultipleActivatedLayers = false;
     }, {
         /**
          * @method moveTo
@@ -297,5 +309,204 @@ Oskari.clazz.define('Oskari.mapframework.domain.Map',
          */
         getSrsName: function () {
             return this._projectionCode;
-        }
+        },
+        /************************************************
+        * Common layer functions
+        ************************************************** */
+        getLayerIndex: function (id, list) {
+            var normalizedId = id + '';
+            if(!list) {
+                list = this.getLayers();
+            }
+            var len = list.length;
+            for (var i = 0; i < len; ++i) {
+                if(list[i].getId() + '' === normalizedId) {
+                    return i;
+                }
+            }
+            return -1;
+        },
+        /************************************************
+        * Selected layers
+        ************************************************** */
+        getLayers : function() {
+            return this._selectedLayers || [];
+        },
+        /**
+         * @public @method getSelectedLayer
+         * Checks if the layer matching the id is added to map
+         *
+         * @param {String} id ID of the layer to check
+         * @return {Oskari.mapframework.domain.AbstractLayer|null} layer or null if not found
+         */
+        getSelectedLayer: function (id) {
+            var index = this.getLayerIndex(id);
+            if(index === -1) {
+                return null;
+            }
+            var list = this.getLayers();
+            return list[index];
+        },
+        /**
+         * @public @method isLayerSelected
+         * Checks if the layer matching the id is added to map
+         *
+         * @param {String} id ID of the layer to check
+         * @return {Boolean} true if the layer is added to map
+         */
+        isLayerSelected: function (id) {
+            return this.getLayerIndex(id) !== -1;
+        },
+        addLayer : function(layer, triggeredBy) {
+            if(!layer || !layer.getId()) {
+                log.warn('Attempt to add layer that is not available.');
+                return;
+            }
+            if(this.isLayerSelected(layer.getId())) {
+                log.warn('Layer already added. Skipping id ' + layer.getId());
+                return false;
+            }
+            this.getLayers().push(layer);
+            var evt = Oskari.eventBuilder('AfterMapLayerAddEvent')(layer);
+            // TODO: setter?
+            evt._creator = triggeredBy;
+            this._sandbox.notifyAll(evt);
+            return true;
+        },
+        removeLayer : function(id, triggeredBy) {
+            var list = this.getLayers();
+            var indexToRemove = this.getLayerIndex(id);
+            if(indexToRemove === -1) {
+                // not found
+                log.debug('Attempt to remove layer "' + id + '" that is not selected.');
+                return false;
+            }
+            // remove from activated list
+            this.deactivateLayer(id);
+            // remove layer
+            var layer = this.getSelectedLayer(id);
+            list.splice(indexToRemove, 1);
+            // notify
+            var evt = Oskari.eventBuilder('AfterMapLayerRemoveEvent')(layer);
+            // TODO: setter?
+            evt._creator = triggeredBy;
+            this._sandbox.notifyAll(evt);
+            return true;
+        },
+        moveLayer : function(id, newIndex, triggeredBy) {
+            var list = this.getLayers();
+            var oldIndex = this.getLayerIndex(id);
+            if(oldIndex === -1) {
+                // no layer to move
+                return false;
+            }
+            if(!Oskari.util.isNumberBetween(newIndex, 0, list.length -1)) {
+                // if not valid index -> treat as "move to last"
+                newIndex = list.length - 1;
+            }
+
+            var moved = Oskari.util.arrayMove(list, oldIndex, newIndex);
+            var layer = this.getSelectedLayer(id);
+            // notify listeners
+            var evt = Oskari.eventBuilder('AfterRearrangeSelectedMapLayerEvent')(layer, oldIndex, newIndex);
+            // TODO: setter?
+            evt._creator = triggeredBy;
+            this._sandbox.notifyAll(evt);
+            return moved;
+        },
+        /************************************************
+        * Activated or "highlighted" layers
+        ************************************************** */
+        getActivatedLayers : function() {
+            return this._activatedLayers || [];
+        },
+        /**
+         * If no parameter is given, returns the state of the flag. With parameter sets the flag.
+         * @param  {Boolean} allow optional boolean value to set the flag
+         * @return {Boolean} value of the flag
+         */
+        allowMultipleActivatedLayers : function(allow) {
+            if(typeof allow === 'undefined') {
+                // getter
+                return this._allowMultipleActivatedLayers;
+            }
+            // setter
+            var newValue = !!allow;
+            var activated = this.getActivatedLayers();
+            // check if we have too many activated layers when turned off
+            if(newValue === false && newValue !== this._allowMultipleActivatedLayers && activated.length > 1) {
+                var latestActivated = activated[activated.length -1];
+                // deactivate all
+                this.deactivateLayer();
+                // reactivate the latest
+                this.activateLayer(latestActivated.getId());
+            }
+            this._allowMultipleActivatedLayers = newValue;
+            return newValue;
+        },
+        /**
+         * @public @method isLayerSelected
+         * Checks if the layer matching the id is added to map
+         *
+         * @param {String} id ID of the layer to check
+         * @return {Boolean} true if the layer is added to map
+         */
+        isLayerActivated: function (id) {
+            return this.getLayerIndex(id, this.getActivatedLayers()) !== -1;
+        },
+        activateLayer : function(id, triggeredBy) {
+            if(!this.isLayerSelected(id)) {
+                log.warn('Trying to activate layer that is not selected. Skipping id ' + id);
+                return false;
+            }
+            if(this.isLayerActivated(layer.getId())) {
+                log.warn('Layer already activated. Skipping id ' + id);
+                return false;
+            }
+            var layer = this.getSelectedLayer(id);
+            // check if multiactivation is allowed -> deactivate the previous if not
+            if(!this.allowMultipleActivatedLayers() && this.getActivatedLayers().length !== 0) {
+                this.deactivateLayer(undefined, triggeredBy);
+            }
+            this.getActivatedLayers().push(layer);
+
+            // finally notify sandbox
+            var evt = Oskari.eventBuilder('map.layer.activation')(layer, true);
+            // TODO: setter?
+            evt._creator = triggeredBy;
+            this._sandbox.notifyAll(evt);
+            return true;
+        },
+        deactivateLayer : function(id, triggeredBy) {
+            var sandbox = this._sandbox;
+            var list = this.getActivatedLayers();
+            var removalList = [];
+            var evtBuilder = Oskari.eventBuilder('map.layer.activation');
+            function notifyDim(removalList) {
+                removalList.forEach(function(layer) {
+                    var evt = evtBuilder(layer, false);
+                    // TODO: setter?
+                    evt._creator = triggeredBy;
+                    sandbox.notifyAll(evt);
+                });
+            }
+            if(typeof id === 'undefined') {
+                // remove all
+                removalList = list.slice(0);
+                this._activatedLayers = [];
+                notifyDim(removalList);
+                return removalList.length !== 0;
+            }
+            // remove single
+            var indexToRemove = this.getLayerIndex(id, list);
+            if(indexToRemove === -1) {
+                // not found
+                log.debug('Attempt to deactivate layer "' + id + '" that is not activated.');
+                return false;
+            }
+            removalList = list.splice(indexToRemove, 1);
+            notifyDim(removalList);
+            return true;
+        },
     });
+}(Oskari));