+++ sources/framework/sandbox/sandbox-map-layer-methods.js
@@ -16,8 +16,7 @@ Oskari.clazz.category('Oskari.Sandbox', 'map-layer-methods', {
      *  layer domain object if found matching id or null if not found
      */
     findMapLayerFromAllAvailable: function (id, name) {
-        var layer = this._core.findMapLayerFromAllAvailable(id, name);
-        return layer;
+        return this.getService('Oskari.mapframework.service.MapLayerService').findMapLayer(id);
     },
 
     /**
@@ -26,9 +25,18 @@ Oskari.clazz.category('Oskari.Sandbox', 'map-layer-methods', {
      * @return {Oskari.mapframework.domain.WmsLayer[]/Oskari.mapframework.domain.WfsLayer[]/Oskari.mapframework.domain.VectorLayer[]/Mixed}
      */
     findAllSelectedMapLayers: function () {
-        var layersList = this._core.getAllSelectedLayers();
+        var layersList = this.getMap().getLayers();
         // copy the array so changing it wont change the core data
         return layersList.slice(0);
+    },
+    /**
+     * @method findAllHighlightedLayers
+     * Returns all currently highlighted map layers
+     * @return {Oskari.mapframework.domain.WmsLayer[]/Oskari.mapframework.domain.WfsLayer[]/Oskari.mapframework.domain.VectorLayer[]/Mixed}
+     */
+    findAllHighlightedLayers: function () {
+        var layersList = this.getMap().getActivatedLayers();
+        return layersList.slice(0);;
     },
 
     /**
@@ -40,8 +48,7 @@ Oskari.clazz.category('Oskari.Sandbox', 'map-layer-methods', {
      *  layer domain object if found matching id or null if not found
      */
     findMapLayerFromSelectedMapLayers: function (layerId) {
-        var layer = this._core.findMapLayerFromSelectedMapLayers(layerId);
-        return layer;
+        return this.getMap().getSelectedLayer(layerId);
     },
 
     /**
@@ -52,18 +59,9 @@ Oskari.clazz.category('Oskari.Sandbox', 'map-layer-methods', {
      * @return {Boolean} true if the layer is added to map
      */
     isLayerAlreadySelected: function (id) {
-        return this._core.isLayerAlreadySelected(id);
+        return this.getMap().isLayerSelected(id);
     },
 
-    /**
-     * @method findAllHighlightedLayers
-     * Returns all currently highlighted map layers
-     * @return {Oskari.mapframework.domain.WmsLayer[]/Oskari.mapframework.domain.WfsLayer[]/Oskari.mapframework.domain.VectorLayer[]/Mixed}
-     */
-    findAllHighlightedLayers: function () {
-        var layer = this._core.getAllHighlightedMapLayers();
-        return layer;
-    },
 
     /**
      * @method isMapLayerHighLighted
@@ -73,14 +71,7 @@ Oskari.clazz.category('Oskari.Sandbox', 'map-layer-methods', {
      * @return {Boolean} true if the layer is highlighted
      */
     isMapLayerHighLighted: function (id) {
-        var highlighted = this.findAllHighlightedLayers(),
-            i;
-        for (i = 0; i < highlighted.length; i++) {
-            if (highlighted[i].getId() + '' === id + '') {
-                return true;
-            }
-        }
-        return false;
+        return this.getMap().isLayerActivated(id);
     },
 
     /**
@@ -90,7 +81,7 @@ Oskari.clazz.category('Oskari.Sandbox', 'map-layer-methods', {
      * @param {Boolean} allow - true to allow, false to restrict to one highlight at a time
      */
     allowMultipleHighlightLayers: function (allow) {
-        this._core.allowMultipleHighlightLayers(allow);
+        this.getMap().allowMultipleActivatedLayers(allow);
     },
 
     /**